+++ opentelemetry-api/src/opentelemetry/context/__init__.py
@@ -12,147 +12,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import threading
-import types
-import typing
-
-try:
-    import contextvars
-except ImportError:
-    contextvars: types.ModuleType = None
+from .base_context import BaseRuntimeContext
 
 __all__ = ['RuntimeContext']
 
+RuntimeContext: BaseRuntimeContext = None
 
-class _RuntimeContext:
-    class Slot:
-        def clear(self):
-            raise NotImplementedError
-
-        def get(self):
-            raise NotImplementedError
-
-        def set(self, value):
-            raise NotImplementedError
-
-    _lock = threading.Lock()
-    _slots: typing.Dict[str, Slot] = {}
-
-    @classmethod
-    def clear(cls):
-        """Clear all slots to their default value."""
-        keys = cls._slots.keys()
-        for name in keys:
-            slot = cls._slots[name]
-            slot.clear()
-
-    @classmethod
-    def register_slot(cls, name, default=None):
-        """Register a context slot with an optional default value.
-
-        :type name: str
-        :param name: The name of the context slot.
-
-        :type default: object
-        :param name: The default value of the slot, can be a value or lambda.
-
-        :returns: The registered slot.
-        """
-        with cls._lock:
-            if name in cls._slots:
-                raise ValueError('slot {} already registered'.format(name))
-            slot = cls.Slot(name, default)
-            cls._slots[name] = slot
-            return slot
-
-    def apply(self, snapshot):
-        """Set the current context from a given snapshot dictionary"""
-
-        for name in snapshot:
-            setattr(self, name, snapshot[name])
-
-    def snapshot(self):
-        """Return a dictionary of current slots by reference."""
-
-        keys = self._slots.keys()
-        return dict((n, self._slots[n].get()) for n in keys)
-
-    def __repr__(self):
-        return '{}({})'.format(type(self).__name__, self.snapshot())
-
-    def __getattr__(self, name):
-        if name not in self._slots:
-            self.register_slot(name, None)
-        slot = self._slots[name]
-        return slot.get()
-
-    def __setattr__(self, name, value):
-        if name not in self._slots:
-            self.register_slot(name, None)
-        slot = self._slots[name]
-        slot.set(value)
-
-    def with_current_context(self, func):
-        """Capture the current context and apply it to the provided func"""
-
-        caller_context = self.snapshot()
-
-        def call_with_current_context(*args, **kwargs):
-            try:
-                backup_context = self.snapshot()
-                self.apply(caller_context)
-                return func(*args, **kwargs)
-            finally:
-                self.apply(backup_context)
-
-        return call_with_current_context
-
-
-class _ThreadLocalRuntimeContext(_RuntimeContext):
-    class Slot(_RuntimeContext.Slot):
-        _thread_local = threading.local()
-
-        def __init__(self, name, default):
-            self.name = name
-            self.default = default if callable(default) else (lambda: default)
-
-        def clear(self):
-            setattr(self._thread_local, self.name, self.default())
-
-        def get(self):
-            try:
-                return getattr(self._thread_local, self.name)
-            except AttributeError:
-                value = self.default()
-                self.set(value)
-                return value
-
-        def set(self, value):
-            setattr(self._thread_local, self.name, value)
-
-
-class _AsyncRuntimeContext(_RuntimeContext):
-    class Slot(_RuntimeContext.Slot):
-        def __init__(self, name, default):
-            self.name = name
-            self.contextvar = contextvars.ContextVar(name)
-            self.default = default if callable(default) else (lambda: default)
-
-        def clear(self):
-            self.contextvar.set(self.default())
-
-        def get(self):
-            try:
-                return self.contextvar.get()
-            except LookupError:
-                value = self.default()
-                self.set(value)
-                return value
-
-        def set(self, value):
-            self.contextvar.set(value)
-
-
-RuntimeContext = _ThreadLocalRuntimeContext()  # pylint:disable=invalid-name
-if contextvars:
-    RuntimeContext = _AsyncRuntimeContext()  # pylint:disable=invalid-name
+try:
+    from .async_context import AsyncRuntimeContext
+    RuntimeContext = AsyncRuntimeContext()
+except ImportError:
+    from .thread_local_context import ThreadLocalRuntimeContext
+    RuntimeContext = ThreadLocalRuntimeContext()
